+++ src/main/java/com/crackedgames/craftics/api/CrafticsAPI.java
@@ -34,7 +34,7 @@ import net.minecraft.item.Item;
  * CrafticsAPI.registerBiome(myBiomeTemplate);
  * CrafticsAPI.registerWeapon(myItem, WeaponEntry.builder(myItem)
  *     .damageType(DamageType.SLASHING).attackPower(8).apCost(1).range(1)
- *     .ability(Abilities.bleed().and(Abilities.sweepAdjacent(0.1, 0.05)))
+ *     .ability(Abilities.sweepAdjacent(0.1, 0.05).and(Abilities.stun(0.05, 0.03)))
  *     .build());
  * }</pre>
  *